refactor(step-8): move note-conventions.ts and its test to src/operations/
Simplest move in the refactor: note-conventions.ts has no cross-file
imports and its test is co-located. Only main.ts's import path needs
updating to ./operations/note-conventions.js.
